fir fonctionnel

+++ fir.h
@@ -1,4 +1,4 @@
 #include "define.h"
 
 absorp firTest(char* filename);
-absorp fir(absorp myAbsorp, int *tab_acr, int *tab_acir);
+absorp fir(absorp myAbsorp, float *tab_acr, float *tab_acir);
